+++ src/CookieCrumble/src/CookieCrumble/TestType.cs
@@ -1,0 +1,8 @@
+namespace CookieCrumble;
+
+internal enum TestType
+{
+    Unknown,
+    Xunit,
+    MSTest
+}
